Create single-use trait to collect HydeKernel Hyperlink calls
